Change short int to int

+++ src/parser.h
@@ -39,7 +39,7 @@
 struct parameter {
   char name[PARSER_MAX_LINE_SIZE];
   char value[PARSER_MAX_LINE_SIZE];
-  short int used;
+  int used;
 };
 
 struct section {
